requirements.txt
python-telegram-bot==20.7
dune-client==1.3.0
python-dotenv==1.0.0
pandas==2.1.4
redis==5.0.1
pytest==7.4.4
pytest-asyncio==0.23.5
prometheus-client==0.19.0
